+++ profile.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-PYTHONPATH=src python3 -m cProfile -s time bin/samovar eg/chairs.samovar --min-events=8000 --unsorted-search
+PYTHONPATH=src python3 -m cProfile -s time bin/samovar eg/chairs.samovar --min-events=8000